Adjusted AWS S3 tests and config

+++ tests/library/CMService/AwsS3Versioning/ClientTest.php
@@ -17,29 +17,30 @@ class CMService_AwsS3Versioning_ClientTest extends CMTest_TestCase {
     public function setUp() {
         $config = CM_Config::get();
         $className = __CLASS__;
+        $version = (string) $config->$className->version;
+        $region = (string) $config->$className->region;
         $key = (string) $config->$className->key;
         $secret = (string) $config->$className->secret;
-        $region = (string) $config->$className->region;
         if (empty($key) || empty($secret)) {
             $this->markTestSkipped('Missing `key` or `secret` config.');
         }
 
-        $this->_client = \Aws\S3\S3Client::factory(array('key' => $key, 'secret' => $secret));
+        $this->_client = new \Aws\S3\S3Client(['version' => $version, 'region' => $region, 'key' => $key, 'secret' => $secret]);
         $this->_client->getConfig()->set('curl.options', array('body_as_string' => true)); // https://github.com/aws/aws-sdk-php/issues/140#issuecomment-25117635
         $this->_bucket = strtolower(str_replace('_', '-', 'test-' . __CLASS__ . uniqid()));
         $this->_filesystem = new CM_File_Filesystem(new CM_File_Filesystem_Adapter_AwsS3($this->_client, $this->_bucket));
 
         $this->_client->createBucket(array('Bucket' => $this->_bucket, 'LocationConstraint' => $region));
-        $this->_client->waitUntilBucketExists(array('Bucket' => $this->_bucket));
         $this->_client->putBucketVersioning(array('Bucket' => $this->_bucket, 'Status' => 'Enabled'));
 
         $this->_restore = new CMService_AwsS3Versioning_Client($this->_client, $this->_bucket, new CM_OutputStream_Null());
     }
 
     public function tearDown() {
-        $clear = new Aws\S3\Model\ClearBucket($this->_client, $this->_bucket);
-        $clear->clear();
-        $this->_client->deleteBucket(array('Bucket' => $this->_bucket));
+        if ($this->_client) {
+            $this->_client->deleteMatchingObjects($this->_bucket, '', '.*');
+            $this->_client->deleteBucket(array('Bucket' => $this->_bucket));
+        }
     }
 
     public function testGetVersioningEnabled() {
